A flag for a valid search param has been created. Updated tab indentation.  - I previously had my vim config setup to have 1 space tabs, but I didn't notice when coding.

+++ pill.sh
@@ -9,32 +9,36 @@ CLR_BLU="$ESC_CHR[34m"
 CLR_MGT="$ESC_CHR[35m"
 
 out_txt_clr() { # takes clr and txt
- echo -e "$1$2$RST_TRM"
+	echo -e "$1$2$RST_TRM"
 }
 
 # this is the actual runtime script
-don=false
+if [[ -z $1 ]]; then
+	out_txt_clr $CLR_RED "No argument found. Requires search parameter."
+	exit 1
+fi
 out_txt_clr $CLR_GRN "Gathering pid(s) associated with \"$1\"..."
 pid=`ps aux | grep $1 | grep -v grep | grep -v bash | awk '{print $2}'`
 shw_pid=`ps aux | grep $1 | grep -v grep | grep -v bash | awk '{printf "%s - %s\n", $2, $11}'`
 if [[ -z $pid ]]; then
- out_txt_clr $CLR_GRN "No processes to kill."
+	out_txt_clr $CLR_GRN "No processes to kill."
 else
- out_txt_clr $CLR_RED "Do you really want to kill processes: \"$CLR_MGT$shw_pid$CLR_RED\"?"
- while [[ $don = false ]]; do
-  out_txt_clr $CLR_YLW "[y/n]"
-  read ans
-  if [[ -n $ans ]]; then
-   don=true
-   if [[ $ans = "y" ]]; then
-    out_txt_clr $CLR_BLU "Killing: \"$CLR_MGT$shw_pid$CLR_BLU\"!"
-    kill $pid
-   else
-    out_txt_clr $CLR_GRN "Alright... have a nice day!"
-   fi
-  else
-   out_txt_clr $CLR_RED "Invalid answer!"
-  fi
- done
+	don=false
+	out_txt_clr $CLR_RED "Do you really want to kill processes: \"$CLR_MGT$shw_pid$CLR_RED\"?"
+ 	while [[ $don = false ]]; do
+  		out_txt_clr $CLR_YLW "[y/n]"
+  		read ans
+  		if [[ -n $ans && $ans = "y" || $ans = "n" ]]; then
+   			don=true
+   			if [[ $ans = "y" ]]; then
+    			out_txt_clr $CLR_BLU "Killing: \"$CLR_MGT$shw_pid$CLR_BLU\"!"
+    			kill $pid
+   			elif [[ $ans = "n" ]]; then
+    			out_txt_clr $CLR_GRN "Alright... have a nice day!"
+   			fi
+  		else
+   			out_txt_clr $CLR_RED "Invalid answer!"
+  		fi
+ 	done
 fi
